Fuck all bug ,解决百度地图计算位置出现的bug，布局填充参数重新设置，fragment返回键退出框论框

+++ app/src/main/java/com/overtech/lenovo/activity/business/tasklist/adapter/TaskListAdapter.java
@@ -119,8 +119,8 @@ public class TaskListAdapter extends Adapter<TaskListAdapter.MyViewHolder> {
             } else {
                 LatLng curLoc = new LatLng(latitude, longitude);
                 LatLng desLoc = new LatLng(task.latitude, task.longitude);
-                double dis = DistanceUtil.getDistance(curLoc, desLoc);
-                holder.taskDistance.setText(dis / 1000 + "km");
+                double dis = DistanceUtil.getDistance(curLoc,desLoc);
+                holder.taskDistance.setText((int)dis / 1000 + "km");
             }
         }
         if (task.appointment_home_mills == 0) {
